Web Scrapping Code/requirements.txt
anyio==3.6.1
async-generator==1.10
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
Flask==2.2.2
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
httptools==0.5.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
pycparser==2.21
pydantic==1.10.2
pymongo==4.2.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.19.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.3
watchfiles==0.17.0
webdriver-manager==3.8.3
websockets==10.3
Werkzeug==2.2.2
wsproto==1.1.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
